// src/app/client-view/models/cash-flow.model.ts
export class CashFlow {
    memberId: string;
    // new api
    houseHoldId: string;

    grossSalary: number;
    investmentIncome: number;
    governmentBenefits: number;
    retirementIncomeStreams: number;
    otherIncome: number;
    totalGrossIncome: number;
    salaryPackaging: number;
    salarySacrifice: number;
    incomeTax: number;
    totalNetIncome: number;
    fixedExpenses: number;
    discretionaryExpenses: number;
    nonDeductibleDebt: number;
    deductibleDebt: number;
    investmentExpenses: number;
    investmentContributions: number;
    personalInsurance: number;
    totalExpenses: number;
    surplusIncome: number;
    weeklySurplus: number;
    primaryClientId: string;

    // new api
    primaryClientFirstName: string;
    primaryClientLastName: string;
    employeeType: number;
}